Make handlers methods of application for DI
This change introduces an `application` struct to house loggers and
other dependencies that many of the handlers will need, and makes
the handlers methods of the said struct. It will make it easier to
unit-test and pass any dependencies we need to handlers in a more
idiomatic, maintainable and scalable way.