Improve the implementation of the template cache
This change automatically parses all pages and all partials using globs. This
way, if new partials or pages are added to their respective directories, no
changes will be necessary and the application will simply start parsing those
new files automatically, so, there is no chance that someone will add a new
template but then forget to add it to some list of known templates that must be
parsed.