Add a common headers middleware to all routes
This change introduces a common headers middleware that is added to all routes
(therefore, we call `commonHeaders(mux)` so `mux` is the `next` parameter for
the middleware function. This way we make sure we add the middleware in such way
that it will always run the middleware before the request hits our servermux.